requirements.txt
celery==4.4.7
ccxt==1.64.51
email-validator==1.1.2
Flask==2.0.2
Flask-JWT-Extended==3.24.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Modals==0.4.1
flask-sock==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
itsdangerous==2.0.1
Jinja2==3.0.2
jinja2-time==0.2.0
PyJWT==1.7.1
PyYAML==5.4
SQLAlchemy==1.4.26
Twisted==18.9.0
Werkzeug==2.0.2
wsproto==1.0.0
WTForms==3.0.0
yarl==1.7.2
python-dotenv==0.19.2
flower==0.9.7
redis==3.5.3
gunicorn==20.1.0
psycopg2-binary==2.9.2
pycoingecko==2.2.0
secure==0.3.0
Flask-Moment==1.0.2
Flask-Executor==0.10.0
sentry-sdk==1.5.1
pycryptodome==3.12.0
scrypt==0.8.19